+++ app/templates/FacturaClie/factura_crud.html
@@ -1,0 +1,70 @@
+{%extends 'index.html'%}
+
+{% block content %}
+<div class='row justify-content-center m-5'>
+    <div class="col-md-10 col-lg-5">
+      <h4><center>Buscar Factura</center></h4>
+      <div  class="form-group">
+        <form id="form_busqueda" name="form_busqueda" method="post">
+            {% csrf_token %}
+          {{ busquedaform.as_p }}  
+          <div  style="float: right;">
+            <button class="btn btn-primary" >Buscar</button>
+            <a type="button" onClick="limpiar()" class="btn btn-warning" >Limpiar</a>
+            <a href="{% url 'aent' %}" type="button" class="btn btn-success" >Agregar</a>
+          </div>                   
+        </form>
+    </div>
+    </div>
+
+<div style="display: flex; justify-content: space-between;">
+    <div>
+        <p>N de entidades encontradas:</p>
+    </div>
+    <div style="display: flex;">
+        <p style="margin:10px;">Mostrados:</p>
+        <select style="width: 80px;" class="form-select form-select-sm" aria-label=".form-select-sm example">
+            <option selected>1-10</option>
+            <option value="1">1</option>
+            <option value="2">2</option>
+            <option value="3">3</option>
+        </select>
+    </div>
+</div>
+<div>
+    <h3 style="text-align: center;">Relación de Entidades</h3>
+</div>
+<table class="table table-striped table-hover">
+    <thead>
+        <tr>
+            <th scope="col">Item</th>
+            <th scope="col">N.Factura</th>
+            <th scope="col">Cliente</th>
+            <th scope="col">Importe</th>
+            <th scope="col">Fecha</th>
+        </tr>        
+    </thead>
+    
+    <tbody>
+        {% if factura_list %}
+        {% for fac in factura_list %}
+        <tr scope="row">
+            <td>{{ forloop.counter }}</td>
+            <td>{{ fac.factura.id }}</td>
+            <td>{{ fac.codcliente.persona.nombre }}</td>
+            <td>{{ fac.factura.totalfactura }}</td>
+            <td>{{ fac.factura.fecha }}</td>
+            <td>
+                <a type="button" href="#" class="btn btn-primary">Ver</a> 
+                <a type="button" href="#" class="btn btn-success">Editar</a>
+                <a type="button" href="#" class="btn btn-danger">Eliminar</a>
+            </td>
+        </tr>
+        {% endfor %}
+        {% else %}
+        <h4>No se encontraron datos</h4>
+        {% endif %}
+    </tbody>
+</table>
+
+{% endblock %}
