+++ app/templates/FacturaClie/facturaclie.html
@@ -1,0 +1,31 @@
+{% extends 'index.html' %}
+
+{% block content %}
+
+    <div class="card container text-center" style="width: 100%; padding-top: 1em; margin-top: 3em; margin-bottom: 3em;">
+        <div class="card-header modal-title fs-5">
+            {{ ent.nombreentidad }}
+        </div>
+        <div class="card-body">
+            <section class="container">
+                <table class="table table-striped">
+                    <thead>
+                    <tr>
+                        <th scope="col">#</th>
+                        <th scope="col">DATOS</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                        <tr>
+                            <th scope="row">NOMBRE:</th>
+                            <td>{{ ent.nombreentidad }}</td>
+                        </tr>
+                    </tbody>
+                </table>
+            </section>
+        </div>
+    </div>
+    <div class="container text-center">
+        <a type="button" href="{% url 'ent' %}" class="btn btn-primary">Regresar</a>
+    </div>
+{% endblock %}
