+++ app/templates/index.html
@@ -47,7 +47,7 @@
                 <a class="nav-link dropdown-toggle nav-link navbar-brand mb-0 h1" href="#" data-bs-toggle="dropdown" aria-expanded="false">Ventas Cliente</a>
                 <ul class="dropdown-menu">
                   <li><a class="dropdown-item" href="{% url 'venta' %}">Venta</a></li>
-                  <li><a class="dropdown-item" href="#">Factura</a></li>
+                  <li><a class="dropdown-item" href="{% url 'facturaclie' %}">Factura</a></li>
                   <li><a class="dropdown-item" href="#">Albaran</a></li>
                   <li><a class="dropdown-item" href="#">Factura Albaran</a></li>
                 </ul>
