+++ app/views/__init__.py
@@ -1,0 +1,12 @@
+from app.views.articulos import *
+from app.views.clientes import *
+from app.views.embalajes import *
+from app.views.entidad import *
+from app.views.familias import *
+from app.views.formas_pago import *
+from app.views.impuestos import *
+from app.views.proveedores import *
+from app.views.ubicaciones import *
+from app.views.ventas_clie import *
+from app.views.factura_clie import *
+
